+++ crates/client/starknet-block-import/Cargo.toml
@@ -1,0 +1,48 @@
+[package]
+name = "mc-starknet-block-import"
+version.workspace = true
+edition.workspace = true
+description = "Starknet specific block import logic"
+homepage = "https://github.com/keep-starknet-strange/madara"
+license = "MIT"
+publish = false
+repository = "https://github.com/keep-starknet-strange/madara"
+
+[package.metadata.docs.rs]
+targets = ["x86_64-unknown-linux-gnu"]
+
+[dependencies]
+async-trait = "0.1"
+log = { workspace = true }
+num-bigint = { workspace = true }
+thiserror = { workspace = true }
+
+# Substrate primitives dependencies
+sp-api = { workspace = true }
+sp-consensus = { workspace = true }
+sp-runtime = { workspace = true }
+
+# Substrate client dependencies
+sc-consensus = { workspace = true }
+sc-rpc-api = { workspace = true }
+
+# Starknet
+blockifier = { workspace = true }
+cairo-lang-starknet = { workspace = true }
+cairo-lang-utils = { workspace = true }
+cairo-vm = { workspace = true }
+hex = { workspace = true }
+madara-runtime = { workspace = true }
+mc-db = { workspace = true }
+mc-rpc = { workspace = true }
+pallet-starknet = { workspace = true }
+pallet-starknet-runtime-api = { workspace = true }
+starknet-core = { workspace = true }
+starknet_api = { workspace = true, features = [
+  "scale-info",
+  "parity-scale-codec",
+] }
+
+# Primitives
+mp-felt = { workspace = true }
+mp-transactions = { workspace = true, features = ["scale-info"] }
