+++ crates/node/Cargo.toml
@@ -27,10 +27,10 @@ async-trait = "0.1"
 clap = { workspace = true, features = ["derive"] }
 futures = { workspace = true, features = ["thread-pool"] }
 log = { workspace = true }
+num-bigint = { workspace = true }
 serde = { workspace = true }
 sha3 = { workspace = true }
 thiserror = { workspace = true }
-num-bigint = { workspace = true }
 
 frame-system = { workspace = true }
 sc-basic-authorship = { workspace = true }
@@ -68,7 +68,6 @@ jsonrpsee = { workspace = true, features = ["server"] }
 sp-api = { workspace = true }
 sp-block-builder = { workspace = true }
 sp-blockchain = { workspace = true }
-sp-consensus = { workspace = true }
 # Substrate client dependencies
 prometheus-endpoint = { workspace = true }
 sc-rpc-api = { workspace = true }
@@ -82,9 +81,6 @@ frame-benchmarking = { workspace = true }
 frame-benchmarking-cli = { workspace = true }
 
 # Starknet
-cairo-vm = { workspace = true }
-cairo-lang-starknet = { workspace = true }
-cairo-lang-utils = { workspace = true }
 blockifier = { workspace = true }
 hex = { workspace = true }
 madara-runtime = { workspace = true }
@@ -96,6 +92,7 @@ mc-l1-messages = { workspace = true }
 mc-mapping-sync = { workspace = true }
 mc-rpc = { workspace = true }
 mc-settlement = { workspace = true, features = ["clap"] }
+mc-starknet-block-import = { workspace = true }
 mc-storage = { workspace = true }
 pallet-starknet = { workspace = true }
 pallet-starknet-runtime-api = { workspace = true }
