+++ starknet-rpc-test/README.md
@@ -1,20 +1,29 @@
 # Starknet RPC tests
 
-Starknet RPC rests are run against a single Madara instance (in order to reduce the CI pipeline time).  
-In order to run tests locally you will need to:
-0. Make sure you have all test dependencies compiled (see next section)
-1. Build Madara binary: `cargo build --profile release`
-2. Setup Madara instance (once): `./target/release/madara --chain=dev --from-local=./configs`
-2. Run Madara node: `./target/release/madara --chain=dev --sealing=manual --alice`
-3. Run tests `cargo test --package starknet-rpc-test works_with_storage_change -- --nocapture`
+Starknet RPC rests are run against a single Madara instance (in order to reduce
+the CI pipeline time).  
+In order to run tests locally you will need to: 0. Make sure you have all test
+dependencies compiled (see next section)
 
-If you need to reset the state, run `./target/release/madara purge-chain --chain=dev`
+1. Build Madara binary: `cargo build --profile release`
+2. Setup Madara instance (once):
+   `./target/release/madara --chain=dev --from-local=./configs`
+3. Run Madara node:
+   `./target/release/madara --chain=dev --sealing=manual --alice`
+4. Run tests
+   `cargo test --package starknet-rpc-test works_with_storage_change -- --nocapture`
+
+If you need to reset the state, run
+`./target/release/madara purge-chain --chain=dev`
 
 ## Generate Cairo contract artifacts
 
-Make sure you have the exact scarb version installed as in `contracts/.tool-versions`. Follow the instructions: https://docs.swmansion.com/scarb/download.html#install-via-installation-script
+Make sure you have the exact scarb version installed as in
+`contracts/.tool-versions`. Follow the instructions:
+<https://docs.swmansion.com/scarb/download.html#install-via-installation-script>
 
-It's necessary to compile the same contract several times with small modifications so that it has different class hash.  
+It's necessary to compile the same contract several times with small
+modifications so that it has different class hash.  
 The artifacts are used for testing the declare class operation.
 
 ```sh
